+++ changelog.md
@@ -4,6 +4,7 @@ TGUI 1.1  (TBD)
 - Added AutoLayout that lets widget fill entire side of parent
 - Added methods for arrow key navigation between widgets
 - Added TextOutlineColor and TextOutlineThickness to ProgressBar renderer
+- Hover state is now reset when mouse leaves the window
 
 
 TGUI 1.0  (30 September 2023)